Replace globals.interrupted with retry_on_fail channel property
Previously, the globals.interrupted variable could be set in order to
request that the channel discontinue retries for failing function calls
(e.g., subscribe and consume). This commit replaces this variable with a
retry_on_fail property which can be set on the client itself in order to
avoid making retries after a failure. The example was updated to set
retry_on_fail to False when the application is interrupted.

This commit also introduces a destroy() method and enter/exit methods
which allow the channel to be used in a with block. When the channel is
exited (or the destroy method is explicitly called), existing resources
are now cleaned up - including an attempt to delete the customer from
the service and cleanup request session resources.