Use proper token if token regenerated
Previously, if a channel call after the initial token was generated were
to return a 401/403 error and a new token was issued, the new token
would be passed in the Auth HTTP header for subsequent requests to the
server without the Realm, "Bearer". In this commit, the proper Realm is
included in requests sent to the server with a regenerated token.
Additionally, the retry logic for token reissuance was simplified
somewhat and the fake consumer service was enhanced with some settings
for triggering periodic token and cookie expirations for testing
purposes.