Specify channel session timeout in seconds instead of milliseconds
Previously, the channel session timeout provided in a channel.create()
call was specified in milliseconds. For consistency with the
`wait_between_queries` parameter on the channel.run() method, this
commit changes the channel session timeout to be specified in seconds.

This commit also includes changes to the fake consumer service which use
the channel session timeout specified on the channel.create() call to
handle when the session expires in the service.