Wrap consumer re-create in run and add stop to channel
Previously, users of the run() method would have to handle
ConsumerErrors raised when a consumer token expires by creating a new
consumer instance and re-subscribing the consumer to the instance in
order to keep consuming records.  This commit expands the run() method
to automatically re-create / re-subscribe a consumer instance in the
event of a token expiration, simplifying the work that a run() method
user would need to do.

This commit also includes a stop() method, which can be invoked in order
to cause an active run() method call to end.