Retry an initial channel produce failure
Previously, if a call to the channel `produce` method were to fail due
to the channel token having expired, no attempt was made to
automatically re-authenticate, attempt to obtain a new token, and repeat
the call to the `produce` method.

With the changes in this commit, an initial failure in a call to the
`produce` method will be retried, with an attempt made to re-authenticate
and obtain a new token before the second attempt to perform the
`produce` call. An error is raised if consecutive calls fail.